Prevent quiet kernel parameter from raising log level.
Prevents the quiet boot option from raising the log level.
It is possible for the loglevel to have previously been set
to something "more quiet" than the quiet value. This occurs
either by default or a preceding  loglevel= option.
This change makes it such that the loglevel= option does
not have to occur after the quiet option to reduce the
loglevel to something below quiet.

See: CLOUD-302 OVER-254 and OVER-2476

I found that its initial commit was:

commit fce43f4cbd1752401d0890fc5412535975ecea85
Author: Sergey Popov <uazure@gmail.com>
Date:   Tue Jan 13 18:37:56 2015 +0200

    CLOUD-302 fix minor kernel bug that allows 'quiet' kernel parameter

    ... to override 'loglevel' kernel parameter with a higher loglevel.