Add quirk for HP dc5750 allow resume from suspend.
When attempting to resume from an S3 (mem) suspend
the machine will start to wake up but fail to
show anything on the monitor. This also appeared
to cause issues with the "freeze" S1 method as well.

Debugging the issue lead to the radeon driver.
There appears to be a common issue for HP machines
with the same radeon chipset "CHIP_RS480".
I added an additional quirk specific to this
machine similar to the existing quirks.
With this change the system is able to resume
normally from suspend.

OVER-2225