drm/i915: add suspend quirk for the Lenovo Thinkpad SL410
The Lenovo Thinkpad SL410 would not successfully complete
the S3 suspend showing a blinking suspend light.
There was an existing quirk which leaves
the device in S0 on hibernation. Extending
the quirk to suspends on the Lenovo Thinkpad SL410
and Lenovo Thinkpad SL510 allowed them to complete
the suspend to ram.

Fixes OVER-1190 and OVER-745.