drm/i915: add suspend quirk for the Lenovo Thinkpad SL410
The Lenovo Thinkpad SL410 would not successfully complete
the S3 suspend showing a blinking suspend light.
There was an existing quirk which would leave
the device in S0 on hibernation. Extending
this quirk to the Lenovo Thinkpad SL410 (and SL510)
in S0 on suspend cases allowed them to suspend.

Fixes OVER-1190 and OVER-745.