Fix for Dell Latitde 3350 ethernet not working.
This patch is related to 18d78b64fddc11eb336f01e46ad3303a3f55d039

It looks like the ethernet card in this system is attempting to use IRQ 18,
which according to 'arch/x86/include/asm/irq_vectors.h' is reserved by the
ACPI subsystem as a system interrupt (everything under 31 is reserved). The
ethernet card seems to share a PCI interrupt (18) with the reserved ACPI
space, which causes the clash.

This patch forces a specific _REV (ACPI revision) value as a workaround
to allow the Dell Latitude 3350's board to work correctly (by causing
Linux to use a different ACPI revision).

OVER-4036