Handle iMac9,1 minimum backlight values.
This extends the nouveau backlight workarounds to include
a minimum value for the brightness. It changes the 0-100
range to be adjusted between he minimum value which is
visible to the maximum value.
Previously the brightness value for the device would
cause the display to be dark for much of the lower end.