CHROMIUM: LSM: fix buffer over-read in printable_cmdline
printable_cmdline() allocates a 4096-byte buffer and copies process
cmdline into that buffer. It then feeds that string into printable().
The string in the buffer may not be null-terminated if the cmdline is
longer than 4096. If that's the case printable() over-reads the buffer
until the next zero.

BUG=b:33159480
TEST=No KASAN errors when running init_module() from a process with
long cmdline

Signed-off-by: Andrey Ulanov <andreyu@google.com>
Change-Id: If9852c0a9d3b46928746288ac1c7547ded6607eb
Reviewed-on: https://chromium-review.googlesource.com/417187
Reviewed-by: Kees Cook <keescook@chromium.org>
(cherry picked from commit 5e358dae2f575165e38090d0a2393192224e3137)