neverware: add TPM whitelist
TPM is a potentially disruptive feature; our previous attempts to
enable it have failed due to runtime bugs (login errors, certificates
not loaded) on some models.

The first models we'd like to certify for TPM support are the Dell
Latitude E5420, E5430, and E6230. The machines we have checked all
have an Atmel TPM chip.

This commit adds the whitelist infrastructure and adds one Atmel
device to the whitelist.

- Add the Atmel TPM vendor ID to the constants in tpm.h

- Add tpm_in_neverware_whitelist function to test whether a device is
  in our whitelist or not

- Update the existing kernel device log to show the vendor ID in
  addition to the device ID so that future devices are easier to
  identify

- During TPM device initialization, use tpm_in_neverware_whitelist to
  disable the driver if the device is not in our whitelist