ALSA: hda - Fix the headphone control order on the Fujitsu Lifebook T731
The Fujitsu Lifebook T731 has two headphones, one is on the Dock
and the other is on the device. Change the def association
such that the built in headphone pin comes before the dock pin.
This change allows the switch control to be associated with
the appropriate pin. Without it the switches would not mute
the appropriate jacks.

OVER-4874