neverware: change TPM whitelist to only affect registration
This fixes sleep/resume on T440p and Yoga 11e G3 when TPM is enabled.

The idea with this change is to allow most of the TPM kernel code to
run normally, but hide the device from userspace. Skipping
registration means that no /sys or /dev files are created for TPM.

https://neverware.atlassian.net/browse/OVER-5816

OVER-5816

Conflicts:
	drivers/char/tpm/tpm-chip.c
	drivers/char/tpm/tpm.h