alsa: Add some workarounds for the 0x14 top speakers pid.
This sets only some verb settings for the various pins
to try to get sound to come out of the top speakers.

This seemed to be some of the changes that came from
the windows driver when it would enable the card.

WIP: The issue here is when the headphones are plugged in
the 0x14 is not muted. It continues to be on.

Perhaps having the control associated with it somehow?

Also, the output seems too low on those top speakers.

Should use a commit which adds some information about the 0x14
as then it might have it's switch exposed.