drm/nouveau/bl: Allocate backlight connector nodes.
Avoid adding nodes to the bl_connectors list from
the stack.

It was being allocated on the stack and added
to the bl_connectors list. nouveau_backlight_exit was
causing a panic when attempting to traverse the bl_connectors
list with nodes from the stack.

This panic occurs on an optimus system when unloading the nouveau
module.

OVER-5961