drm/nouveau/fbcon: Fix NULL pointer access in nouveau_fbcon_destroy.
It is possible for this to get called with a null helper.fb.
This can cause a panic when unloading the nouveau module on
a Optimus system.

Upstream master commit 34112bf4935d ("drm/nouveau/fbcon: fix NULL
pointer access in nouveau_fbcon_destroy") includes part of the fix of
this patch.

OVER-5961