Add quirk for Dell Optiplex 5060
This is to make the rear audio port to work with the headset.
Used the "options snd-hda-intel mdoel=dell-headset-dock"in the
alsa.conf initially to see if it works. Later, added the
quirk in the kernel itself to avoid putting the quirk in the
alsa.conf. This is fixed at the pulseaudio layer for systems
using pulseaudio. It would be best to fix this in the kernel.

OVER-7489