Revert "firmware: dmi_scan: Use lowercase letters for UUID"
This reverts commit 712ff25450bd01366301eef81c33e865d901e7b7.

The product_uuid can be used by cryptohome as the system key, which
makes it harder to remove the hard drive and decrypt it on a different
machine. The UUID used to be uppercase, but upstream switched it to
lowercase.

Unfortunately the kernel maintainer doesn't care that he broke a
public kernel ABI:
https://lists.gt.net/linux/kernel/3171307

We could also fix this problem in cryptohome, but it feels a little
safer to do it in the kernel in case there are other places I'm not
aware of that also depend on the UUID.

https://neverware.atlassian.net/browse/OVER-9798