Add default config for beerover64.
This is one that I use to put a debug kernel onto a device.
It is loosely built on the beerover64 config file with
some known useful debug options I use.

To build a kernel with this config run the steps.
```
make O=../build-out beerover64-debug_defconfig
cd ../build-out
make all
```