Revert "ANDROID: net: wireless: Decrease scan entry expiration to avoid stall results"
This reverts commit a116b28a5364b439306692bdada504cf35ee0c14.

It was found that wifi scans do not complete when attempting
to connect to hidden wifi access points. This was particularly
noticeable on the BCM4321 [14e4:4328] b43 devices.

This change is NOT in the linux kernel just the CrOS kernel
and came along with the rest of the CrOS changes when
4.19 was merged. See OVER-8694 for the merge.

I tested connecting to a hidden network on an iMac7,1 with the
BCM4321 [14e4:4328] wifi device using the b43 driver.

This fixes OVER-10030