Revert "Enable cold boot attack mitigation"
This reverts commit 23611f752de3f1558a81656e8f2a453091208893.

This was causing some machines to not boot. It should be safe to
revert because it is already implemented in
ccc829ba3624beb9a703fc995d016b836d9eead8, "efi/libstub: Enable reset
attack mitigation". The implementation details of that commit are
slightly different, and it must be enabled with
`CONFIG_RESET_ATTACK_MITIGATION=y`. Doing so does not seem to cause
any problems with booting.

https://neverware.atlassian.net/browse/OVER-10239