Disable EFI Runtime Services for all ProBook 11 G1 EE.
It is fair to overshoot this check instead of undershoot.
The device that was tested matches
"HP ProBook x360 11 G1 EE" exactly.
This change will disable EFI runtime services on machines
that we are not seeing a problem like the Pentium
version as they share the same product name.

There are many different SKUs available for these machines.
Disable EFI runtime on all of them to avoid the particular
problem.

OVER-9289

Conflicts:
	init/main.c