drm/nouveau/devinit: Force post for Apple MCP79 devices.
Apple iMac 9,1 has flickering issues when an external
monitor is attached.
Forcing a post on these devices fixes the flickering
and allows the monitor to be used.
Limit the quirk to only Apple models that were tested
for the issue.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=88272
OVER-10385