[NEVERWARE/UPSTREAM] ALSA: hda: Modify stream stripe mask only when needed
The recent in HD-audio stream management for changing the stripe
control seems causing a regression on some platforms.  The stripe
control is currently used only by HDMI codec, and applying the stripe
mask unconditionally may lead to scratchy and static noises as seen on
some MacBooks.

For addressing the regression, this patch changes the stream
management code to apply the stripe mask conditionally only when the
codec driver requested.

OVER-10928

Note: This patch is not in the kernel officially yet. It was provided by
Takashi Iwai in the Bugzilla linked below.

Fixes: 9b6f7e7a296e ("ALSA: hda: program stripe bits for controller")
BugLink: https://bugzilla.kernel.org/show_bug.cgi?id=204477
Cc: <stable@vger.kernel.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>