neverware: quirk some more Dell Optiplex models to fix headphone out
We already had quirks for some Dell Optiplex models to use
`ALC269_FIXUP_DELL2_MIC_NO_PRESENCE`. This isn't actually the right
fix; it gets headphone output working but not mic input. Also it
incorrectly adds a dock headphone jack, but it's only visible in
alsamixer or similer, not from the UI.

I added a comment to make it clearer that these quirks are hacky, and
four more Optiplex models to the list.

OVER-5958