Fallthrough I915 X_TILED format
Skip this error check if we're using the X tiled format. On kernels 4.19
and earlier this check passed but the code here has since been
refactored and fails causing Intel machines to flicker.

This should probably be investigated and fixed at the browser level in
the future but for now this works and matches the 4.19 behavior.