drm/i915: Avoid an invalid deref of i915_mch_dev.
Do not dereference the i915_mch_dev pointer if it
is NULL.

It is possible for intel-ips to call i915_read_mch_val
before the i915 driver has initialized i915_mch_dev.

This causes a panic from ips-monitor on a machine that has
an intel_ips device, the i915 module built in and does not
have an intel graphics device.

OVER-11721