Check HTS_GTD_DIS for the initial gpu turbo setting.
update_turbo_limits will check the absence of HTS_GTD_DIS
to determine the gpu_turbo_enabled flag. It will only
do this after the symbols in ips_get_i915_syms are loaded.
Move the update_turbo_limits call to follow symbol linking
so that it can detect the initial gpu enabled state for
the device.

The device for which I tested (T510) this change has the bit cleared
so it is effectively the same result.