Assume turbo is off when gpu_lower fails.
The i915_gpu_lower fails only if the i915 device is not
yet registered. In this case any future attempt to disable
turbo for i915 will also fail as the device isn't registered.
Assume that if there is no device that it shouldn't attempt
to disable turbo on the device. This is indicated by changing
the __gpu_turbo_on value to false.

i915_gpu_turbo_disable could potentially fail because of the
card being busy so it isn't just checked there.
This change avoids a problem where every 5 seconds when the ips-adjust
process loops it fails to disable and logs the failure indefinitely

This was seen on a device which has intel_ips but doesn't have an
intel graphics device. The symbols are able to load as i915 is
built-in to the kernel that is being used.

OVER-12425

(cherry picked from commit 9e5bfcc3030cbb99e41220667955ce64d1a6aa01)