neverware: add an allow list for devices that can use SOF
We want to prevent a big sudden QA burden, so limit SOF devices based
on system DMI info.

The allow list can be bypassed by adding this to kernel params:

    snd_intel_dspcfg.sof_bypass_allow_list=1

Note: originally I tried putting this DMI check at the top of
`sof_pci_probe`, but it didn't work correctly in the disallow case;
the output audio device would not show up in that case. Switched it to
`snd_intel_dsp_driver_probe` instead where it seems to work fine.

[OVER-13135]