neverware: Revert "BACKPORT: gpio: gpiolib: Allow GPIO IRQs to lazy disable"
This reverts commit 8257a4bbfec325366af2abb55d7fdd7324f61374.

---

This gpio change was causing the ASUS T100TA to hang on boot, see Jira
issue for serial log.

I verified that the backport commit looks just like the original with
some minor naming changes, and looked a bit in upstream master to see
if I could spot a fixup commit that might work instead of reverting,
but didn't find anything.

[OVER-13292]