neverware: fix TTY switching
Switching from the browser to a frecon VT was failing because the
browser could not call successfully call `drmDropMaster`. CrOS has a
special hack that should allow the browser to make that call even
though it doesn't run as root:
https://chromium.googlesource.com/chromiumos/third_party/kernel/+/13ac39b4037510c4fa294d061f8d154d0ccd8d38

This call was being disallowed by the lockdown LSM. Arguably the right
fix here is to move the `drm_master_relax` file to a better location
under sysfs, but that would require larger changes. Instead, just add
a hardcoded exception for this path so that the lockdown doesn't apply
to it.

[OVER-13234]