BACKPORT: FROMGIT: mac80211: drop A-MSDUs on old ciphers
With old ciphers (WEP and TKIP) we shouldn't be using A-MSDUs
since A-MSDUs are only supported if we know that they are, and
the only practical way for that is HT support which doesn't
support old ciphers.

However, we would normally accept them anyway. Since we check
the MMIC before deaggregating A-MSDUs, and the A-MSDU bit in
the QoS header is not protected in TKIP (or WEP), this enables
attacks similar to CVE-2020-24588. To prevent that, drop A-MSDUs
completely with old ciphers.

Cc: stable@vger.kernel.org
Link: https://lore.kernel.org/r/20210511200110.076543300172.I548e6e71f1ee9cad4b9a37bf212ae7db723587aa@changeid
Signed-off-by: Johannes Berg <johannes.berg@intel.com>

(cherry picked from commit 270032a2a9c4535799736142e1e7c413ca7b836e
 git://git.kernel.org/pub/scm/linux/kernel/git/jberg/mac80211.git master)

Conflicts:
   net/mac80211/rx.c
   [Fixed conflict in copyright header]

BUG=b:187853103
TEST=Verified wifi_matfunc/perf

Signed-off-by: Abhishek Kumar <kuabhs@chromium.org>
Change-Id: I2f150e6d42ec9e5dca6eb2712ceb2cf610f41930
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2890113
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
(cherry picked from commit eee40ac6bf9f83eb294d6109f95ed54920269574)